+++ main.html
@@ -1,0 +1,42 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="test.css">
+    <script src="https://kit.fontawesome.com/b9cf35b5bd.js"></script>
+    <link rel="preconnect" href="https://fonts.gstatic.com">
+    <link href="https://fonts.googleapis.com/css2?family=Monda&display=swap" rel="stylesheet">
+    <title>Document</title>
+</head>
+
+<body>
+
+  <form id="diet_form" onSubmit={onSubmit}>
+    <h4>Choose the value which best describes your meat consumption ...</h4>
+    <p><input onChange={onChange} type="radio" name="meat_level" id="meat_full" value="1000" />
+        <label for="meat_full">I eat meat 7 days a week.</label>
+    </p>
+    <p>
+        <input onChange={onChange} type="radio" name="meat_level" id="meat_partial" value="600" />
+        <label for="meat_partial">I eat meat 4 - 6 days per week.</label>
+    </p>
+    <p>
+        <input onChange={onChange} type="radio" name="meat_level" id="meat_low" value="400" />
+        <label for="meat_low">I eat meat 1 - 3 days per week.</label>
+    </p>
+    <p>
+        <input onChange={onChange} type="radio" name="meat_level" id="meat_none" value="200" />
+        <label for="meat_none">I don't eat meat.</label>
+    </p>
+
+    <input class="button" type="submit" value="Calculate" name="submit" />        
+
+</form>
+
+
+</body>
+
+</html>
